Refactor getReports : utilisation de offset et limit en query string
- offset et limit en optionnels
- offset en Long et limit en Int
- offset ne peut pas être négatif. Pas de contrôle sur un offset qui serait supérieur au nombre total de signalements
- limit doit être inférieur à 250 pour être pris en compte. 25 par défaut

+++ app/repositories/ReportRepository.scala
@@ -112,7 +112,7 @@ class ReportRepository @Inject()(dbConfigProvider: DatabaseConfigProvider)(impli
         .result
     )
 
-  def getReports(offset: Int, limit: Int): Future[PaginatedResult[Report]] = db.run {
+  def getReports(offset: Long, limit: Int): Future[PaginatedResult[Report]] = db.run {
       for {
         reports <- reportTableQuery
           .sortBy(_.creationDate)
